0.2 simple_robot Twist_To_Motors interface implemented by applying differential drive classic equations. moveMotors-plugin changed from Int to Float to simplify the control (standrd input from the keypboard)

+++ simple_robot_control/scripts/mock_velocities.py
@@ -1,18 +1,18 @@
 #!/usr/bin/env python
 
 import rospy
-from std_msgs.msg import Int32
+from std_msgs.msg import Float32
 
 def mock_velocoties():
-    pub_left_vel = rospy.Publisher('lWheels', Int32, queue_size=5)
-    pub_right_vel = rospy.Publisher('rWheels', Int32, queue_size=5)
+    pub_left_vel = rospy.Publisher('lWheels', Float32, queue_size=5)
+    pub_right_vel = rospy.Publisher('rWheels', Float32, queue_size=5)
 
     rospy.init_node('mock_velocities', anonymous=True)
     rate = rospy.Rate(20)
 
     while not rospy.is_shutdown():
-        vl = Int32(1)
-        vr = Int32(1)
+        vl = Float32(1)
+        vr = Float32(1)
         #rospy.loginfo("Publishing mock velocities to topics...")
         pub_left_vel.publish(vl)
         pub_right_vel.publish(vr)
